backend/pyproject.toml
[project]
name = "ica-backend"
version = "0.1.0"
description = "Intelligent Coding Agent Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "qdrant-client>=1.7.0",
    "celery[redis]>=5.3.6",
    "alembic>=1.13.1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]
